+++ examples/simulate_fmi2_cs.c
@@ -24,7 +24,7 @@ int main(int argc, char* argv[]) {
     CALL(FMI2EnterInitializationMode(S));
     CALL(FMI2ExitInitializationMode(S));
 
-    for (size_t step = 0;; step++) {
+    for (uint64_t step = 0;; step++) {
 
         CALL(recordVariables(S, outputFile));
